feat: add frontend form validation and UX improvements

+++ frontend/src/app/pages/list-items/list-items.ts
@@ -3,6 +3,7 @@ import { Component, inject, OnInit, signal } from '@angular/core';
 import { Item } from '../../models/item.model';
 import { ItemService } from '../../services/item.service';
 import { RouterLink } from "@angular/router";
+import { ToastService } from '../../services/toast.service';
 
 
 @Component({
@@ -16,6 +17,9 @@ export class ListItems implements OnInit{
 
   // inyectamos el servicio, acceso backend CRUD
   private itemService = inject(ItemService);
+
+  // inyectamos servicio Toast
+  private toastService = inject(ToastService);
 
   // estado reactivo, cuando items cambia Angular actualiza automáticamente, array items
   items = signal<Item[]> ([]);
@@ -32,6 +36,8 @@ export class ListItems implements OnInit{
       },
       // si hay error
       error: (error: any) => {
+        // mensaje toast
+        this.toastService.show('Failed to load items', 'error');
         console.error(error);
       }
     });
@@ -39,6 +45,14 @@ export class ListItems implements OnInit{
 
   // eliminar item concreto
   deleteItem(id: number) {
+    // confirmamos eliminar
+    const confirmed = confirm(
+      'Are you sure you want to dlete this item?'
+    );
+    // no confirmado, salir del método, no llamar al backend
+    if (!confirmed) {
+      return;
+    }
     // llamada delete al backend
     this.itemService.deleteItem(id).subscribe({
       // si responde bien
@@ -46,9 +60,13 @@ export class ListItems implements OnInit{
         console.log('Item eliminado');
         // actualizar lista items, crea nuevo array sin el item eliminado
         this.items.update(items => items.filter(item => item.id! !== id));
+        // mensaje toast
+        this.toastService.show('Item deleted succesfully', 'success');
       },
       // si hay error
       error: (error: any) => {
+        // mensaje toast
+        this.toastService.show('Failed to delete item', 'error');
         console.error(error);
       }
